1/requirements.txt
openai-whisper
torch
fastapi
uvicorn
python-multipart